garden_scorer/requirements.txt

requests
flask
sqlalchemy==1.2.10
PyMySQL==0.7.11
tornado
jaeger-client
opentracing_instrumentation
Flask-Opentracing
flask-cors
